+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * FEATURE: Add `to_json` presenter method (@boblail)
+* FEATURE: Add `to_csv` presenter method (@boblail)
 
 ## v0.4.0 (2019 Apr 16)